+++ nv/shell_windows.py
@@ -19,6 +19,9 @@ import os
 import subprocess
 import sys
 import tempfile
+import traceback
+
+from NeoVintageous.nv.options import get_option
 
 # https://mypy.readthedocs.io/en/latest/common_issues.html#python-version-and-system-platform-checks
 assert sys.platform.startswith('win')
@@ -26,14 +29,11 @@ assert sys.platform.startswith('win')
 try:
     import ctypes
 except ImportError:
-    from sublime import platform
-    if platform() == 'windows':
-        raise EnvironmentError('ctypes module missing for Windows.')
-
+    traceback.print_exc()
     ctypes = None
 
 
-def get_startup_info():
+def _get_startup_info():
     # Hide the child process window.
     startupinfo = subprocess.STARTUPINFO()
     startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
@@ -41,45 +41,62 @@ def get_startup_info():
     return startupinfo
 
 
-def get_oem_cp() -> str:
+def _get_encoding() -> str:
     return str(ctypes.windll.kernel32.GetOEMCP())
 
 
-def run_and_wait(view, cmd) -> None:
-    subprocess.Popen(['cmd.exe', '/c', cmd + '&& pause']).wait()
+# TODO Review Do newlines really need to converted on Windows?
+def _translate_newlines(text: str):
+    return text.replace('\r\n', '\n')
 
 
-def run_and_read(view, cmd) -> str:
-    out, err = subprocess.Popen(['cmd.exe', '/c', cmd],
-                                stdout=subprocess.PIPE,
-                                stderr=subprocess.PIPE,
-                                shell=True,
-                                startupinfo=get_startup_info()).communicate()
-
-    try:
-        return (out or err).decode(get_oem_cp()).replace('\r\n', '\n')
-    except AttributeError:
-        return ''
+def open(view) -> None:
+    term = get_option(view, 'term')
+    if term:
+        subprocess.Popen([term, '/k'], cwd=os.getcwd())
 
 
-def filter_region(view, txt, command) -> str:
+def read(view, cmd: str) -> str:
+    p = subprocess.Popen([get_option(view, 'shell'), '/c', cmd],
+                         stdout=subprocess.PIPE,
+                         stderr=subprocess.PIPE,
+                         startupinfo=_get_startup_info())
+
+    out, err = p.communicate()
+
+    if out:
+        return _translate_newlines(out.decode(_get_encoding()))
+
+    if err:
+        return _translate_newlines(err.decode(_get_encoding()))
+
+    return ''
+
+
+def filter_region(view, txt: str, cmd: str) -> str:
     try:
         contents = tempfile.NamedTemporaryFile(suffix='.txt', delete=False)
         contents.write(txt.encode('utf-8'))
         contents.close()
 
         script = tempfile.NamedTemporaryFile(suffix='.bat', delete=False)
-        script.write(('@echo off\ntype %s | %s' % (contents.name, command)).encode('utf-8'))
+        script.write(('@echo off\ntype %s | %s' % (contents.name, cmd)).encode('utf-8'))
         script.close()
 
         p = subprocess.Popen([script.name],
                              stdout=subprocess.PIPE,
                              stderr=subprocess.PIPE,
-                             startupinfo=get_startup_info())
+                             startupinfo=_get_startup_info())
 
         out, err = p.communicate()
 
-        return (out or err).decode(get_oem_cp()).replace('\r\n', '\n')[:-1].strip()
+        if out:
+            return _translate_newlines(out.decode(_get_encoding()))
+
+        if err:
+            return _translate_newlines(err.decode(_get_encoding()))
+
+        return ''
     finally:
         os.remove(script.name)
         os.remove(contents.name)